+++ crates/scouter_server/tests/api/mod.rs
@@ -1,0 +1,4 @@
+pub mod drift;
+pub mod health;
+pub mod profile;
+pub mod user;
