+++ crates/scouter_server/tests/lib.rs
@@ -1,0 +1,2 @@
+pub mod api;
+pub mod common;
